backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-decouple
requests
psycopg2-binary
django-redis
